# Go micro performance. lexer atn config using murmur hash 32.