--- a/runtime/Go/antlr/dfa_state.go
+++ b/runtime/Go/antlr/dfa_state.go
@@ -92,7 +92,7 @@ func NewDFAState(stateNumber int, configs ATNConfigSet) *DFAState {
 
 // GetAltSet gets the set of all alts mentioned by all ATN configurations in d.
 func (d *DFAState) GetAltSet() *Set {
-	alts := NewSet(nil, nil)
+	alts := NewSet(nil, nil, nil)
 
 	if d.configs != nil {
 		for _, c := range d.configs.GetItems() {