The Go target needs some TLC at both a micro and macro level. A a micro level the lexer (and probably the parser) is slow, around 40x slower than a hand rolled one. Where the cpp target is only around 40% slower (see https://github.com/antlr/antlr4/issues/1540 & https://github.com/pboyer/antlr4/issues/92). At a macro level the LR optimisation (see https://github.com/antlr/antlr4/issues/192) have not been ported to the Go runtime.

This PR changed the LexerATNConfig hash code from a string that was hashed with fnv to hashcode ints that are murmur32 hashed. This gives about a 30% (benchmark number below) improvement (if implemented correctly), so there is still low hanging fruit in the 40x slow down. I intend to revisit this when I have some time or performance becomes an issue. My currently environment doesn't support profiling, which is probably an effective way of finding micro issues. Anyone interested in helping with this, I would be more than happy to collaborate.

@parrt and @sharwell any gotchas in implementing the LR changes?

**Comments**
**Benchmarking** - At this point in Antlr4's maturity it would be nice to have a benchmarking framework to complement  the unit testing framework. Go's benchmarking is an extension of testing and by the looks of it was influenced by [caliper](https://github.com/google/caliper), a Java benchmarking framework. Any suggestion of similar tools for the other targets and what such a framework might look like? Do any travis add-ins support reporting on benchmark data?

**Code coverage** - Whilst looking at the Go and Java code I came across another micro issue. In the Java runtime the `boolean add(T t)` in `Array2DHashSet` class will always return true (ie new element added), this behaviour also exists in the Go runtime. This doesn't effect behaviour, but without benchmarking it is difficult to tell the benefit from fixing it or replacing it with a hash set from a thirdparty library (eg [guava](https://github.com/google/guava) r19 is java 1.7 compatible). It wouldn't be a surprise if many micro issues exist across the runtime code bases. In this case the code in question dates back to 2012. So a deeper question is how to find them. The first answer that springs to mind is code coverage. Has anyone setup code coverage on Antlr4 using one of the travis add-ins? What other tools and techniques might be of interest?

Cheers
Gary

Go benchmarking numbers for this PR (see [modified dgraph benchmark](https://github.com/millergarym/dgraph/blob/bench-antlr4/antlr4go/lexer/benchmark_test.go))
@ashishne thought this might interest you.
before
```
bash-3.2$ go test -test.run=XXX -v  -benchtime=2s -bench . -cpuprofile cprof.prof
BenchmarkQuery/q1.1-12             10000            398395 ns/op
BenchmarkQuery/q1.2-12             10000            432751 ns/op
BenchmarkQuery/q2.1-12             10000            446644 ns/op
BenchmarkQuery/q2.2-12              5000            514398 ns/op
BenchmarkQuery/q3.1-12              5000            490114 ns/op
BenchmarkQuery/q3.2-12              5000            504340 ns/op
BenchmarkQuery/q4.1-12              5000            537611 ns/op
BenchmarkQuery/q4.2-12              5000            552284 ns/op
BenchmarkQuery/q5.1-12              5000            538648 ns/op
BenchmarkQuery/q5.2-12              5000            552259 ns/op
```

after
```
bash-3.2$ go test -test.run=XXX -v  -benchtime=2s -bench .
BenchmarkQuery/q1.1-12             10000            277898 ns/op
BenchmarkQuery/q1.2-12             10000            299240 ns/op
BenchmarkQuery/q2.1-12             10000            307816 ns/op
BenchmarkQuery/q2.2-12             10000            351049 ns/op
BenchmarkQuery/q3.1-12             10000            336241 ns/op
BenchmarkQuery/q3.2-12             10000            352299 ns/op
BenchmarkQuery/q4.1-12             10000            374569 ns/op
BenchmarkQuery/q4.2-12             10000            384309 ns/op
BenchmarkQuery/q5.1-12             10000            374470 ns/op
BenchmarkQuery/q5.2-12             10000            386929 ns/op
```

From https://github.com/antlr/antlr4/issues/1540
```
query$ gotb -test.run=XXX -benchtime=2s -v
BenchmarkQueryParse/spielberg:handwitten:-4               100000         46918 ns/op
BenchmarkQueryParse/spielberg:antlr:-4                      2000       1741364 ns/op
BenchmarkQueryParse/tomhanks:handwitten:-4                100000         25982 ns/op
BenchmarkQueryParse/tomhanks:antlr:-4                       2000       1654579 ns/op
BenchmarkQueryParse/nestedquery:handwritten:-4             30000         73053 ns/op
BenchmarkQueryParse/nestedquery:antlr:-4                    1000       3385005 ns/op
PASS
ok      github.com/dgraph-io/dgraph/query    21.922s
```