@parrt 2017-02-23
verrrry interesting. Can Go people from @antlr/antlr-targets take a look at this?