@willfaught 2017-02-24
Are both commits related? One is so small, and the other so large, it'd be nice if they could be squashed if so.

Can we simplify the "Go micro performance. Changed lexer atn config from Hash() string to …
…HashCode() int. Using murmur hash 32." commit summary to something like: "Change lexer ATN config hash to murmur 32"?