@millergarym 2017-02-24
If possible can some one with hashing / crypto experience to do a high level look at this. I'm not at all confident at the approach I've taken.

@parrt any suggestions as to who is expert in this area?