@davesisson 2017-02-24
I second the import an existing implementation approach.  To get you started I found this particular implementation (there's a murmur hash in there) but haven't vetted it for capability:

https://github.com/leemcloughlin/gofarmhash