@millergarym 2017-02-24
@willfaught thanks for the feedback, much appreciated. This PR was more about generating discussing than being merged and I've learnt a lot from your comments. Hopefully some hashing expert will provide some high level feedback. Time permitting I'll revisit this.

Any comments on including benchmarking as a standard part of the build?

Cheers
Gary