@pboyer 2017-02-25
@millergarym Thanks, this is a good start. Do you plan to follow up on this? If not, I could probably start a new PR incorporating some of these ideas. You're right that there's a bunch of performance improvement to be done on the Go runtime.