@millergarym 2017-02-26
@pboyer, (cc: @parrt, @willfaught) lets close this PR and start a new one. I'm happy to help and think the best use of my time is looking into automated benchmarking. Travis doesn't seem to support this, however Jenkins does or can be made to via a combination of plugins. I'm happy to look into a Docker based Jenkins solution for non Travis CI stuff (benchmarking, cover coverage, todos and maybe even releases eg the XPath cyclic dependencies issue https://github.com/antlr/antlr4/issues/1620). Thoughts?